Use inline for C++ instead of unnamed namespace
Using an unnamed namespace meant every file that used these functions
got a different version of them, which leads to One-Definition Rule
violations if called by templates. Marking them inline means every file
gets the same definition, and the linker discards duplicates at
link-time.

The constexpr specifier implies inline, so inline is not needed when
FASTMOD_API is defined to constexpr.

A different macro is needed for the constexpr variables in the C++
function templates, so that they aren't marked 'inline' when that's what
FASTMOD_API expands to.